split in_controller to DSL, common runtime, and rails specific logic
