+++ lib/declarative_authorization.rb
@@ -1,5 +1,8 @@
 require File.join(%w{declarative_authorization helper})
-require File.join(%w{declarative_authorization in_controller})
+if defined?(ActionController)
+  require File.dirname(__FILE__) + '/declarative_authorization/controller/rails.rb'
+end
+
 if defined?(ActiveRecord)
   require File.join(%w{declarative_authorization in_model})
   require File.join(%w{declarative_authorization obligation_scope})
@@ -12,7 +15,9 @@ end
 
 require File.join(%w{declarative_authorization railsengine}) if defined?(::Rails::Engine)
 
-ActionController::Base.send :include, Authorization::AuthorizationInController
-ActionController::Base.helper Authorization::AuthorizationHelper
+if defined?(ActionController)
+  ActionController::Base.send :include, Authorization::Controller::Rails
+  ActionController::Base.helper Authorization::AuthorizationHelper
+end
 
 ActiveRecord::Base.send :include, Authorization::AuthorizationInModel if defined?(ActiveRecord)
