+++ lib/declarative_authorization/controller_permission.rb
@@ -58,7 +58,7 @@ module Authorization
                 "#{load_object_model} from params[:id] " +
                 "(#{contr.params[:id].inspect}), because attribute_check is enabled " +
                 "and #{instance_var.to_s} isn't set, but failed: #{e.class.name}: #{e}")
-            raise if AuthorizationInController.failed_auto_loading_is_not_found?
+            raise if Authorization::Controller::Runtime.failed_auto_loading_is_not_found?
           end
           contr.instance_variable_set(instance_var, object)
         end
