+++ test/controller_test.rb
@@ -266,10 +266,10 @@ class LoadObjectControllerTest < ActionController::TestCase
       request!(MockUser.new(:test_role), "show", reader)
     end
 
-    Authorization::AuthorizationInController.failed_auto_loading_is_not_found = false
+    Authorization::Controller::Runtime.failed_auto_loading_is_not_found = false
     request!(MockUser.new(:test_role), "show", reader)
     assert !@controller.authorized?
-    Authorization::AuthorizationInController.failed_auto_loading_is_not_found = true
+    Authorization::Controller::Runtime.failed_auto_loading_is_not_found = true
   end
 
   def test_filter_access_with_object_load_custom
